Reorganize code for NVDA add-on
